move Component

+++ app/app/src/main/java/com/app/remote_controller_app/components/Button.java
@@ -23,7 +23,7 @@ public class Button extends Component implements OutputComponent {
     @JsonCreator
     public Button(@JsonProperty("name") String name, @JsonProperty("id") String id,
                   @JsonProperty("sizeX") int sizeX, @JsonProperty("sizeY") int sizeY,
-                  @JsonProperty("posX") float posX, @JsonProperty("posY") float posY,
+                  @JsonProperty("posX") int posX, @JsonProperty("posY") int posY,
                   @JsonProperty("msg") String msg){
         super(name, id, sizeX, sizeY, posX, posY);
         this.msg = msg;
