+++ app/app/src/main/java/com/app/remote_controller_app/components/LED.java
@@ -28,7 +28,7 @@ public class LED extends Component implements InputComponent{
     @JsonCreator
     public LED(@JsonProperty("name") String name, @JsonProperty("id") String id,
                @JsonProperty("sizeX") int sizeX, @JsonProperty("sizeY") int sizeY,
-               @JsonProperty("posX") float posX, @JsonProperty("posY") float posY){
+               @JsonProperty("posX") int posX, @JsonProperty("posY") int posY){
         super(name, id, sizeX, sizeY, posX, posY);
         is_on = false;
     }
